LayoutTests/platform/gtk/accessibility/aria-list-and-listitem.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div role="list">
<div role="listitem" aria-label="item 1"><img src='#' width=100 height=100></div>
<div role="listitem" aria-label="item 2"><img src='#' width=100 height=100></div>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the ARIA roles of list and listitem map correctly to Mac accessibility roles.");

    if (window.accessibilityController) {

          // this text field should be required.
          document.getElementById("body").focus();
          var obj = accessibilityController.focusedElement;

          shouldBe("obj.childAtIndex(0).role", "'AXRole: list'");
          shouldBe("obj.childAtIndex(0).childAtIndex(0).role", "'AXRole: list item'");
          shouldBe("obj.childAtIndex(0).childAtIndex(1).role", "'AXRole: list item'");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
