LayoutTests/platform/gtk/accessibility/aria-list-and-listitem-expected.txt



This tests that the ARIA roles of list and listitem map correctly to Mac accessibility roles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.childAtIndex(0).role is 'AXRole: list'
PASS obj.childAtIndex(0).childAtIndex(0).role is 'AXRole: list item'
PASS obj.childAtIndex(0).childAtIndex(1).role is 'AXRole: list item'
PASS successfullyParsed is true

TEST COMPLETE
